+++ src/core/settings/GOSettingFile.cpp
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2024 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -19,7 +19,7 @@ GOSettingFile::GOSettingFile(
 wxString GOSettingFile::Validate(const wxString &value) const {
   wxString newValue = value;
 
-  if (newValue == wxEmptyString || !wxFileExists(newValue))
+  if (newValue.IsEmpty())
     newValue = GetDefaultValue();
   if (!newValue.IsEmpty()) {
     // convert value to an absolute path