+++ src/grandorgue/GOMetronome.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -45,6 +45,8 @@ private:
   GOLabelControl m_MeasureDisplay;
   GORank *m_rank;
   unsigned m_StopID;
+
+  const wxString &GetSampleFilePath(bool isFirst);
 
   void HandleTimer() override;