+++ src/grandorgue/gui/dialogs/settings/GOSettingsOptions.h
@@ -1,6 +1,6 @@
 /*
  * Copyright 2006 Milan Digital Audio LLC
- * Copyright 2009-2025 GrandOrgue contributors (see AUTHORS)
+ * Copyright 2009-2026 GrandOrgue contributors (see AUTHORS)
  * License GPL-2.0 or later
  * (https://www.gnu.org/licenses/old-licenses/gpl-2.0.html).
  */
@@ -15,7 +15,6 @@ template <class T> class GOChoice;
 class GOConfig;
 class wxCheckBox;
 class wxChoice;
-class wxDirPickerCtrl;
 class wxSpinCtrl;
 
 class GOSettingsOptions : public wxPanel {
@@ -42,8 +41,6 @@ class GOSettingsOptions : public wxPanel {
     ID_RECORD_DOWNMIX,
     ID_VOLUME,
     ID_LANGUAGE,
-    ID_METRONOME_MEASURE,
-    ID_METRONOME_BPM,
     ID_NEW_BAS_MEL,
   };
 
@@ -73,8 +70,6 @@ private:
   wxChoice *m_Interpolation;
   wxSpinCtrl *m_MemoryLimit;
   wxChoice *m_Language;
-  wxSpinCtrl *m_MetronomeMeasure;
-  wxSpinCtrl *m_MetronomeBPM;
   wxCheckBox *m_CheckForUpdatesAtStartup;
 
   wxString m_OldLanguageCode;